Item collection changes
Option buttons now aren't created when the player already have the item (item state is 3).
Now the text of the narration are basic, and the item descriptions are added if the item is not in player's possession. (item state not 3).
Also changed the 'func', having the 'chk' to check if the item's state is 3, because using the 'func' automatically changed it's state to 3.
Also implemented the Notification when the player picks the item, with the popup fading in and out.